First shot at actually moving code.
